requirements.txt
python-dotenv==1.0.0
aiogram==3.13.1
openai
telegramify-markdown
python-telegram-bot
